- Ajuste do atributo Estoque para int

+++ APICatalogo/Domain/Produto.cs
@@ -24,7 +24,7 @@ namespace APICatalogo.Domain
         [Required]
         [StringLength(300)]
         public string? ImagemUrl { get; set; }
-        public float Estoque { get; set; }
+        public int Estoque { get; set; }
         public DateTime DataCadastro { get; set; }
 
         public int CategoriaId { get; set; }
